adding the monitoring task
