+++ locksmith/src/worker/worker.ts
@@ -11,9 +11,11 @@ import { addKeyJobs } from './tasks/addKeyJobs'
 import { addHookJobs } from './tasks/hooks/addHookJobs'
 import { sendHook } from './tasks/hooks/sendHook'
 import { sendEmail } from './tasks/sendEmail'
+import { monitor } from './tasks/monitor'
 import { Pool } from 'pg'
 
 const crontabProduction = `
+*/5 * * * * monitor
 */5 * * * * addRenewalJobs
 0 0 * * * addRenewalJobsDaily
 0 0 * * 0 addRenewalJobsWeekly
@@ -22,6 +24,7 @@ const crontabProduction = `
 `
 
 const cronTabTesting = `
+*/1 * * * * monitor
 */1 * * * * addRenewalJobs
 0 0 * * * addRenewalJobsDaily
 0 0 * * * addRenewalJobsWeekly
@@ -38,11 +41,13 @@ export async function startWorker() {
       // @ts-expect-error - type is not defined properly
       ssl: config.database?.dialectOptions?.ssl,
     }),
+    // logger: plug into our main logger
     crontab,
     concurrency: 5,
     noHandleSignals: false,
     pollInterval: 1000,
     taskList: {
+      monitor,
       addRenewalJobs,
       addRenewalJobsDaily,
       addRenewalJobsWeekly,
